pyproject.toml
[project]
name = "api-django"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0",
    "django-filter>=25.2",
    "djangorestframework>=3.16.1",
    "djangorestframework-xml>=2.0.0",
    "djoser>=2.3.3",
    "markdown>=3.10",
    "pytest>=9.0.2",
    "pytest-django>=4.11.1",
]
